test(sfrmaker/utils.py): wrap doctest in float to pass regardless of numpy 1 or 2

+++ sfrmaker/utils.py
@@ -351,7 +351,7 @@ def width_from_arbolate_sum(asum, a=0.1193, b=0.5032, minimum_width=1., input_un
     --------
     Original equation from Feinstein et al (2010), for arbolate sum of 1,000 km:
     >>> width = width_from_arbolate_sum(1000, 0.1193, 0.5032, input_units='kilometers', output_units='feet')
-    >>> round(width, 2)
+    >>> float(round(width, 2))
     124.69
     """
     if not np.isscalar(asum):
